# internal/delivery/http/docs/swagger.yaml
definitions:
...
  models.CreateIngredientRequest:
    properties:
      count:
        type: number
      dimension:
        type: string
      name:
        type: string
      recipe_id:
        type: string
    type: object
  models.CreateIngredientResponse:
    properties:
      guid:
        type: string
    type: object
...
  models.CreateStepRequest:
    properties:
      cooking_time:
        type: number
      description:
        type: string
      order_number:
        type: integer
      recipe_id:
        type: string
    type: object
  models.CreateStepResponse:
    properties:
      guid:
        type: string
    type: object
...
  models.DeleteIngredientResponse:
    properties:
      success:
        type: boolean
    type: object
...
  models.DeleteStepResponse:
    properties:
      success:
        type: boolean
    type: object
  models.ErrorBadRequest:
    properties:
      error_code:
        type: integer
      message:
        type: string
    type: object
...
  models.Ingridient:
    properties:
      dimension:
        type: string
      guid:
        type: string
      name:
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      access_token_ttl:
        type: integer
      refresh_token:
        type: string
      refresh_token_ttl:
        type: integer
    type: object
...
  models.RegistrationRequest:
    properties:
      password:
        type: string
      retry_password:
        type: string
      username:
        type: string
    type: object
  models.RegistrationResponse:
    properties:
      guid:
        type: string
    type: object
  models.UpdateIngredientRequest:
    properties:
      count:
        type: number
      dimension:
        type: string
      guid:
        type: string
      name:
        type: string
      recipe_id:
        type: string
    type: object
  models.UpdateIngredientResponse:
    properties:
      success:
        type: boolean
    type: object
...
  models.UpdateStepRequest:
    properties:
      cooking_time:
        type: number
      description:
        type: string
      guid:
        type: string
      order_number:
        type: integer
      recipe_id:
        type: string
    type: object
  models.UpdateStepResponse:
    properties:
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Registration user
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Registration user
      tags:
      - auth
  /v1/auth/registration:
    post:
      consumes:
      - application/json
      description: Registration user
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.RegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegistrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Registration user
      tags:
      - auth
  /v1/cooking-step:
    post:
      consumes:
      - application/json
      description: Create cooking step
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateStepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateStepResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Create cooking step
      tags:
      - cooking-step
    put:
      consumes:
      - application/json
      description: Update cooking step
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateStepResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Update cooking step
      tags:
      - cooking-step
  /v1/cooking-step/{id}:
    delete:
      consumes:
      - application/json
      description: Delete cooking step by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteStepResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Delete cooking step by id
      tags:
      - cooking-step
...
  /v1/recipe-ingridient:
    delete:
      consumes:
      - application/json
      description: Delete ingridient by id
      parameters:
      - description: recipe_id
        in: query
        name: recipe_id
        required: true
        type: string
      - description: ingredient_id
        in: query
        name: ingredient_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeleteIngredientResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Delete ingridient by id
      tags:
      - recipe-ingridient
    get:
      consumes:
      - application/json
      description: ingridient list
      parameters:
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ingridient'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      summary: ingridient list
      tags:
      - recipe-ingridient
    post:
      consumes:
      - application/json
      description: Create ingridient
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateIngredientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreateIngredientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Create ingridient
      tags:
      - recipe-ingridient
    put:
      consumes:
      - application/json
      description: Update ingridient
      parameters:
      - description: data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateIngredientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateIngredientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorBadRequest'
      security:
      - ApiKeyAuth: []
      summary: Update ingridient
      tags:
      - recipe-ingridient
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"